Add AudioManager for general audio configuration
Now the AudioManager class can be accessed to call a sound from
any script.

Every sound now can be configurated inside the AudioManager class,
with the clip of the sound, volume, pitch, loop, and add a
mixer source to it.

The sound volume can be ajusted by the user in the Main Menu.

A prefab for the AudioManager was created for his use in any scene.

Most of the script was developed following the tutorial from
Brackeys, in the link:
https://www.youtube.com/watch?v=6OT43pvUyfY