Refactor the HUD script
The code before was used to test the hud reposition accordingly
to the screen size, but in every frame of the game.

Now, the calculation of the position of the two elements of the HUD
represented by the THUNDER and the POINTS is calculated only at the
start of the scene.